Numerics moved to ircclient package
